requirements.txt
numpy
opencv-python
autopy
matplotlib
seaborn
Pillow
pyscreenshot
PyAutoGUI